Improve makefile, add simple MPI test
